hope

+++ 0x00-hello_world/6-size.c
@@ -8,9 +8,9 @@
 int main(void)
 {
 	printf("Size of a char: %d byte(s)\n", sizeof(char));
-	printf("Size of an int: %i byte(s)\n", sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(long));
-	printf("Size of a long long int: %zu bytes(s)\n", (unsigned long)sizeof(long long));
-	printf("Size of a float: %zu bytes(s)\n", (unsigned long)sizeof(float));
+	printf("Size of an int: %d byte(s)\n", sizeof(int));
+	printf("Size of a long int: %d byte(s)\n", sizeof(long));
+	printf("Size of a long long int: %d bytes(s)\n", sizeof(long long));
+	printf("Size of a float: %d bytes(s)\n", sizeof(float));
 	return (0);
 }
